Refactor SignedInLinks & added PropTypes
